+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.5.0
+* [Server] Added `disable-legacy-query-handler`
+* [Client] Reconfigure some initial option values to be less annoying
+
 # 1.4.0
 * Replaced default implementation of `UserNameToIdResolver` (`usercache.json`) with a better alternative that is more efficient and prevents cache corruption
 * Fix NPE when a new user intially joins a server